wip prefer client TaskConfig over driver state's
The driver is persisting and migrating a TaskConfig which will break
everything if used because it will contain paths and alloc ids for the
old alloc.

Still not quite working: the replacement alloc gets restarted
immediately. Either the client is killing the wrong task on migration,
or my fakeremote driver isn't properly splitting the old/new handle.